Reworked shared filters handling [#4536].
