+++ lib/ui/messages/MessagesDirectoryPanel.dart
@@ -2,6 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/ui/messages/MessagesConversationPanel.dart';
 import 'package:illinois/ui/messages/MessagesWidgets.dart';
+import 'package:illinois/ui/profile/ProfileDirectoryAccountsList.dart';
 import 'package:illinois/ui/profile/ProfileDirectoryAccountsPage.dart';
 import 'package:illinois/ui/profile/ProfileDirectoryPage.dart';
 import 'package:illinois/ui/profile/ProfileDirectoryWidgets.dart';
@@ -11,6 +12,7 @@ import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/model/auth2.directory.dart';
 import 'package:rokwire_plugin/model/social.dart';
 import 'package:rokwire_plugin/service/localization.dart';
+import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/social.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
@@ -26,12 +28,10 @@ class MessagesDirectoryPanel extends StatefulWidget {
   _MessagesDirectoryPanelState createState() => _MessagesDirectoryPanelState();
 }
 
-class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with TickerProviderStateMixin {
+class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with TickerProviderStateMixin implements NotificationsListener {
   final GlobalKey<RecentConversationsPageState> _recentPageKey = GlobalKey();
-  final GlobalKey<ProfileDirectoryAccountsPageState> _allUsersPageKey = GlobalKey();
-  final GlobalKey<ConversationSearchBarState> _searchBarKey = GlobalKey();
+  GlobalKey<ProfileDirectoryAccountsPageState> _allUsersPageKey = GlobalKey();
   final ScrollController _recentScrollController = ScrollController();
-  final ConversationsSearchController _searchController = ConversationsSearchController();
   final ScrollController _allUsersScrollController = ScrollController();
 
   late TabController _tabController;
@@ -42,10 +42,18 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
     Localization().getStringEx('panel.messages.directory.tab.all.label', 'All Users'),
   ];
 
+  String _searchText = '';
+  Map<String, dynamic> _filterAttributes = <String, dynamic>{};
+
   final Set<String> _selectedAccountIds = <String>{};
 
   @override
   void initState() {
+    NotificationService().subscribe(this, [
+      Social.notifyConversationsUpdated,
+      Social.notifyMessageSent,
+    ]);
+
     _tabController = TabController(length: _tabNames.length, initialIndex: _selectedTab, vsync: this);
     _tabController.addListener(_onTabChanged);
 
@@ -54,12 +62,30 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
 
   @override
   void dispose() {
+    NotificationService().unsubscribe(this);
+
     _tabController.removeListener(_onTabChanged);
     _tabController.dispose();
+
     _recentScrollController.dispose();
     _allUsersScrollController.dispose();
 
     super.dispose();
+  }
+
+  // NotificationsListener
+  @override
+  void onNotification(String name, dynamic param) {
+    if (name == Social.notifyConversationsUpdated) {
+      if (mounted) {
+        setState(() {
+          _selectedAccountIds.clear();
+        });
+      }
+    }
+    else if (name == Social.notifyMessageSent) {
+      setState(() {});
+    }
   }
 
   @override
@@ -83,7 +109,7 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
       ),
       Padding(
         padding: const EdgeInsets.symmetric(horizontal: 16.0),
-        child: ConversationsSearchBar(key: _searchBarKey, searchController: _searchController, showFilters: _selectedTab != 0,),
+        child: _searchBarWidget,
       ),
       Padding(
         padding: const EdgeInsets.symmetric(horizontal: 16.0),
@@ -124,6 +150,7 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
     return tabs;
   }
 
+
   Widget _buildContinueButtonOverlay(Widget content, { ScrollController? scrollController }) {
     return RefreshIndicator(onRefresh: _onRefresh, child:
       Stack(
@@ -158,8 +185,7 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
   Widget get _recentContent =>
     RecentConversationsPage(
       key: _recentPageKey,
-      searchController: _searchController,
-      initialSearch: searchText,
+      searchText: _searchText,
       recentConversations: widget.recentConversations,
       conversationPageSize: widget.conversationPageSize,
       scrollController: _recentScrollController,
@@ -168,15 +194,42 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
     );
 
   Widget get _allUsersContent =>
-    ProfileDirectoryAccountsPage(DirectoryAccounts.userDirectory,
-      displayMode: DirectoryDisplayMode.select,
+    ProfileDirectoryAccountsList(DirectoryAccounts.userDirectory,
       key: _allUsersPageKey,
+      displayMode: DirectoryDisplayMode.select,
       scrollController: _allUsersScrollController,
-      searchController: _searchController,
-      initialSearch: searchText,
+      searchText: _searchText,
+      filterAttributes: _filterAttributes,
       onAccountSelectionChanged: _onAccountSelectionChanged,
       selectedAccountIds: _selectedAccountIds,
     );
+
+  // Search & Filters
+
+  Widget get _searchBarWidget =>
+    ProfileDirectoryFilterBar(
+      key: ValueKey(ProfileDirectoryFilter(searchText: _searchText, attributes: _filterAttributes)),
+      searchText: _searchText,
+      onSearchText: _onSearchText,
+      filterAttributes: (_selectedTab != 0) ? _filterAttributes : null,
+      onFilterAttributes: _onFilterAttributes,
+    );
+
+  void _onSearchText(String text) {
+    setStateIfMounted((){
+      _searchText = text;
+      _allUsersPageKey = GlobalKey();
+    });
+  }
+
+  void _onFilterAttributes(Map<String, dynamic> filterAttributes) {
+    setStateIfMounted((){
+      _filterAttributes = filterAttributes;
+      _allUsersPageKey = GlobalKey();
+    });
+  }
+
+  // Event Handlers
 
   Future<void> _onTapCreateConversation() async {
     // do not need to check for existing conversation with selected members because Social BB handles it
@@ -198,12 +251,6 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
       setState(() {
         _selectedTab = _tabController.index;
       });
-
-      if (_tabController.index == 0) {
-        _recentPageKey.currentState?.onConversationsTabChanged(searchText, filterAttributes);
-      } else {
-        _allUsersPageKey.currentState?.onConversationsTabChanged(searchText, filterAttributes);
-      }
     }
   }
 
@@ -240,8 +287,6 @@ class _MessagesDirectoryPanelState extends State<MessagesDirectoryPanel> with Ti
 
   bool get _hasSelectedAccounts => _selectedAccountIds.isNotEmpty;
 
-  String get searchText => _searchBarKey.currentState?.searchText ?? '';
-  Map<String, dynamic> get filterAttributes => _searchBarKey.currentState?.filterAttributes ?? {};
 }
 
 enum MessagesDirectoryContentType { recent, all }
