+++ lib/ui/messages/MessagesWidgets.dart
@@ -1,15 +1,8 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/service/Analytics.dart';
-import 'package:illinois/ui/attributes/ContentAttributesPanel.dart';
 import 'package:illinois/ui/profile/ProfileDirectoryWidgets.dart';
 import 'package:illinois/utils/AppUtils.dart';
-import 'package:rokwire_plugin/model/content_attributes.dart';
-import 'package:rokwire_plugin/model/group.dart';
 import 'package:rokwire_plugin/model/social.dart';
-import 'package:rokwire_plugin/service/auth2.dart';
-import 'package:rokwire_plugin/service/auth2.directory.dart';
-import 'package:rokwire_plugin/service/groups.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/social.dart';
@@ -17,16 +10,16 @@ import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class RecentConversationsPage extends StatefulWidget {
-  final ConversationsSearchController searchController;
-  final String? initialSearch;
+  final String? searchText;
   final ScrollController? scrollController;
   final List<Conversation> recentConversations;
   final int conversationPageSize;
   final void Function(bool, Conversation)? onConversationSelectionChanged;
   final Set<String>? selectedAccountIds;
 
-  RecentConversationsPage({super.key, required this.searchController, required this.recentConversations, required this.conversationPageSize,
-    this.initialSearch, this.scrollController, this.onConversationSelectionChanged, this.selectedAccountIds });
+  RecentConversationsPage({super.key, required this.recentConversations, required this.conversationPageSize,
+    this.searchText, this.scrollController,
+    this.onConversationSelectionChanged, this.selectedAccountIds });
 
   @override
   State<StatefulWidget> createState() => RecentConversationsPageState();
@@ -42,20 +35,15 @@ class RecentConversationsPageState extends State<RecentConversationsPage> with A
   bool _canExtend = false;
 
   String? _expandedConversationId;
-  Set<String> _selectedIds = <String>{};
-  String? _searchText;
 
   // Map<String, dynamic> _filterAttributes = <String, dynamic>{};
 
   @override
   void initState() {
     NotificationService().subscribe(this, [
-      Social.notifyConversationsUpdated,
       Social.notifyMessageSent,
     ]);
 
-    widget.searchController.onUpdateSearchText = _onSearchConversations;
-    _searchText = widget.initialSearch;
     widget.scrollController?.addListener(_scrollListener);
     _conversations = widget.recentConversations;
     _conversationsMap = _buildConversationsMap(widget.recentConversations);
@@ -65,8 +53,8 @@ class RecentConversationsPageState extends State<RecentConversationsPage> with A
 
   @override
   void dispose() {
-    widget.scrollController?.removeListener(_scrollListener);
     NotificationService().unsubscribe(this);
+    widget.scrollController?.removeListener(_scrollListener);
     super.dispose();
   }
 
@@ -76,45 +64,29 @@ class RecentConversationsPageState extends State<RecentConversationsPage> with A
   // NotificationsListener
   @override
   void onNotification(String name, dynamic param) {
-    if (name == Social.notifyConversationsUpdated) {
-      if (mounted) {
-        setState(() {
-          _selectedIds.clear();
-        });
-        _load();
-      }
-    }
-    else if (name == Social.notifyMessageSent) {
+    if (name == Social.notifyMessageSent) {
       if (mounted) {
         _load();
       }
     }
-  }
-
-  @override
-  void didUpdateWidget(covariant RecentConversationsPage oldWidget) {
-    super.didUpdateWidget(oldWidget);
-    _selectedIds = widget.selectedAccountIds ?? {};
   }
 
   @override
   Widget build(BuildContext context) {
     super.build(context);
 
-    List<Widget> contentList = <Widget>[];
     if (_loadingProgress) {
-      contentList.add(_loadingContent);
+      return _loadingContent;
     }
     else if (_conversations == null) {
-      contentList.add(_messageContent(Localization().getStringEx('panel.messages.directory.conversations.failed.text', 'Failed to load recent conversations.')));
+      return _messageContent(Localization().getStringEx('panel.messages.directory.conversations.failed.text', 'Failed to load recent conversations.'));
     }
     else if (_conversations?.isEmpty == true) {
-      contentList.add(_messageContent(Localization().getStringEx('panel.messages.directory.conversations.empty.text', 'There are no recent conversations.')));
+      return _messageContent(Localization().getStringEx('panel.messages.directory.conversations.empty.text', 'There are no recent conversations.'));
     }
     else {
-      contentList.add(_conversationsContent);
+      return _conversationsContent;
     }
-    return Column(crossAxisAlignment: CrossAxisAlignment.start, children: contentList);
   }
 
   Widget get _conversationsContent {
@@ -126,7 +98,7 @@ class RecentConversationsPageState extends State<RecentConversationsPage> with A
         contentList.add(RecentConversationCard(conversation,
           expanded: (_expandedConversationId != null) && (conversation.id == _expandedConversationId),
           onToggleExpanded: conversation.isGroupConversation ? () => _onToggleConversationExpanded(conversation) : null,
-          selected: ListUtils.contains(_selectedIds, conversation.memberIds, checkAll: true) ?? false, // conversation must contain all selected account IDs to be selected
+          selected: ListUtils.contains(widget.selectedAccountIds, conversation.memberIds, checkAll: true) ?? false, // conversation must contain all selected account IDs to be selected
           onToggleSelected: (value) => widget.onConversationSelectionChanged?.call(value, conversation),
         ));
       }
@@ -137,18 +109,6 @@ class RecentConversationsPageState extends State<RecentConversationsPage> with A
     }
 
     return Column(crossAxisAlignment: CrossAxisAlignment.start, children: contentList);
-  }
-
-  void onConversationsTabChanged(String searchText, Map<String, dynamic> _) {
-    widget.searchController.onUpdateSearchText = _onSearchConversations;
-    _onSearchConversations(searchText);
-  }
-
-  void _onSearchConversations(String searchText) {
-    if (_searchText != searchText) {
-      _searchText = searchText;
-      _load();
-    }
   }
 
   void _onToggleConversationExpanded(Conversation conversation) {
@@ -197,7 +157,7 @@ class RecentConversationsPageState extends State<RecentConversationsPage> with A
       });
 
       //TODO: add time intervals to filter
-      List<Conversation>? conversations = await Social().loadConversations(offset: 0, limit: widget.conversationPageSize, name: _searchText);
+      List<Conversation>? conversations = await Social().loadConversations(offset: 0, limit: widget.conversationPageSize, name: widget.searchText);
       setStateIfMounted(() {
         _loading = false;
         _loadingProgress = false;
@@ -225,7 +185,7 @@ class RecentConversationsPageState extends State<RecentConversationsPage> with A
       });
 
       //TODO: add time intervals to filter
-      List<Conversation>? conversations = await Social().loadConversations(offset: _conversationsCount, limit: widget.conversationPageSize, name: _searchText);
+      List<Conversation>? conversations = await Social().loadConversations(offset: _conversationsCount, limit: widget.conversationPageSize, name: widget.searchText);
       if (mounted && _extending && !_loading) {
         setState(() {
           if (conversations != null) {
@@ -260,201 +220,6 @@ class RecentConversationsPageState extends State<RecentConversationsPage> with A
     }
     return map;
   }
-}
-
-class ConversationsSearchController {
-  Function(String)? onUpdateSearchText;
-  Function(Map<String, dynamic>)? onUpdateFilterAttributes;
-
-  ConversationsSearchController({ this.onUpdateSearchText, this.onUpdateFilterAttributes });
-}
-
-class ConversationsSearchBar extends StatefulWidget {
-  final ConversationsSearchController searchController;
-  final bool showFilters;
-
-  ConversationsSearchBar({ super.key, required this.searchController, this.showFilters = true });
-
-  @override
-  State<StatefulWidget> createState() => ConversationSearchBarState();
-}
-
-class ConversationSearchBarState extends State<ConversationsSearchBar> {
-
-  String _searchText = '';
-  final TextEditingController _searchTextController = TextEditingController();
-  final FocusNode _searchFocusNode = FocusNode();
-  Map<String, dynamic> _filterAttributes = <String, dynamic>{};
-
-  @override
-  void dispose() {
-    _searchTextController.dispose();
-    _searchFocusNode.dispose();
-
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(padding: const EdgeInsets.only(bottom: 16), child:
-      Row(children: [
-        Expanded(child:
-          Container(decoration: _searchBarDecoration, padding: EdgeInsets.only(left: 16), child:
-            Row(children: <Widget>[
-              Expanded(child:
-                _searchTextWidget
-              ),
-              _searchImageButton('close',
-                label: Localization().getStringEx('panel.search.button.clear.title', 'Clear'),
-                hint: Localization().getStringEx('panel.search.button.clear.hint', ''),
-                rightPadding: _searchImageButtonHorzPadding / 2,
-                onTap: _onTapClear,
-              ),
-              _searchImageButton('search',
-                label: Localization().getStringEx('panel.search.button.search.title', 'Search'),
-                hint: Localization().getStringEx('panel.search.button.search.hint', ''),
-                leftPadding: _searchImageButtonHorzPadding / 2,
-                onTap: _onTapSearch,
-              ),
-            ],)
-          )
-        ),
-        if (widget.showFilters)
-          Padding(padding: EdgeInsets.only(left: 6), child:
-          _filtersButton
-          ),
-      ],),
-    );
-  }
-
-  Widget get _searchTextWidget =>
-      Semantics(
-          label: Localization().getStringEx('panel.messages.directory.conversations.search.field.label', 'SEARCH FIELD'),
-          hint: null,
-          textField: true,
-          excludeSemantics: true,
-          value: _searchTextController.text,
-          child: TextField(
-            controller: _searchTextController,
-            focusNode: _searchFocusNode,
-            decoration: InputDecoration(border: InputBorder.none,),
-            style: Styles().textStyles.getTextStyle('widget.input_field.dark.text.regular.thin'),
-            cursorColor: Styles().colors.fillColorSecondary,
-            keyboardType: TextInputType.text,
-            autocorrect: false,
-            autofocus: false,
-            maxLines: 1,
-            onSubmitted: (_) => _onTapSearch(),
-          )
-      );
-
-  Decoration get _searchBarDecoration => BoxDecoration(
-    color: Styles().colors.white,
-    border: Border.all(color: Styles().colors.disabledTextColor, width: 1),
-    borderRadius: BorderRadius.circular(12),
-  );
-
-  Widget _searchImageButton(String image, { String? label, String? hint, double leftPadding = _searchImageButtonHorzPadding, double rightPadding = _searchImageButtonHorzPadding, void Function()? onTap }) =>
-      Semantics(label: label, hint: hint, button: true, excludeSemantics: true, child:
-        InkWell(onTap: onTap, child:
-          Padding(padding: EdgeInsets.only(left: leftPadding, right: rightPadding, top: _searchImageButtonVertPadding, bottom: _searchImageButtonVertPadding), child:
-            Styles().images.getImage(image, excludeFromSemantics: true),
-          ),
-        ),
-      );
-
-  static const double _searchImageButtonHorzPadding = 16;
-  static const double _searchImageButtonVertPadding = 12;
-
-  Widget get _filtersButton =>
-      InkWell(onTap: _onFilter, child:
-        Container(decoration: _searchBarDecoration, padding: EdgeInsets.symmetric(vertical: 14, horizontal: 14), child:
-          Styles().images.getImage('filters') ?? SizedBox(width: 18, height: 18,),
-        ),
-      );
-
-  void _onFilter() {
-    Analytics().logSelect(target: 'Filters');
-    ContentAttributes? directoryAttributes = _directoryAttributes;
-    if (directoryAttributes != null) {
-      Navigator.push(context, CupertinoPageRoute(builder: (context) => ContentAttributesPanel(
-        title: Localization().getStringEx('panel.profile.directory.accounts.filters.header.title', 'Directory Filters'),
-        description: Localization().getStringEx('panel.profile.directory.accounts.filters.header.description', 'Choose at leasrt one attribute to filter the User Directory.'),
-        scope: Auh2Directory.attributesScope,
-        contentAttributes: directoryAttributes,
-        selection: _filterAttributes,
-        sortType: ContentAttributesSortType.alphabetical,
-        filtersMode: true,
-      ))).then((selection) {
-        if ((selection != null) && mounted) {
-          setState(() {
-            _filterAttributes = selection;
-          });
-          widget.searchController.onUpdateFilterAttributes?.call(_filterAttributes);
-        }
-      });
-    }
-  }
-
-  ContentAttributes? get _directoryAttributes {
-    ContentAttributes? directoryAttributes = Auth2().directoryAttributes;
-    if (directoryAttributes != null) {
-      ContentAttribute? groupAttribute = _groupAttribute;
-      if (groupAttribute != null) {
-        directoryAttributes = ContentAttributes.fromOther(directoryAttributes);
-        directoryAttributes?.attributes?.add(groupAttribute);
-      }
-      return directoryAttributes;
-    }
-    else {
-      return null;
-    }
-  }
-
-  static const String _groupAttributeId = 'group';
-
-  ContentAttribute? get _groupAttribute {
-    List<Group>? userGroups = Groups().userGroups;
-    return ((userGroups != null) && userGroups.isNotEmpty) ?
-    ContentAttribute(
-        id: _groupAttributeId,
-        title: Localization().getStringEx('panel.profile.directory.accounts.attributes.event_type.hint.empty', 'My Groups'),
-        emptyHint: Localization().getStringEx('panel.profile.directory.accounts.attributes.event_type.hint.empty', 'Select groups'),
-        semanticsHint: Localization().getStringEx('panel.profile.directory.accounts.home.attributes.event_type.hint.semantics', 'Double type to show groups.'),
-        widget: ContentAttributeWidget.dropdown,
-        scope: <String>{ Auh2Directory.attributesScope },
-        requirements: null,
-        values: List.from(userGroups.map<ContentAttributeValue>((Group group) => ContentAttributeValue(
-          label: group.title,
-          value: group.id,
-        )))
-    ) : null;
-  }
-
-  void _onTapClear() {
-    Analytics().logSelect(target: 'Search Clear');
-    if (_searchText.isNotEmpty) {
-      setState(() {
-        _searchTextController.text = _searchText = '';
-      });
-      _searchFocusNode.unfocus();
-      widget.searchController.onUpdateSearchText?.call(_searchText);
-    }
-  }
-
-  void _onTapSearch() {
-    Analytics().logSelect(target: 'Search Text');
-    if (_searchText != _searchTextController.text) {
-      setState(() {
-        _searchText = _searchTextController.text;
-      });
-      _searchFocusNode.unfocus();
-      widget.searchController.onUpdateSearchText?.call(_searchText);
-    }
-  }
-
-  String get searchText => _searchText;
-  Map<String, dynamic> get filterAttributes => _filterAttributes;
 }
 
 class RecentConversationCard extends StatefulWidget {
