+++ lib/ui/profile/ProfileDirectoryAccountsList.dart
@@ -1,0 +1,277 @@
+
+import 'dart:math';
+
+import 'package:flutter/material.dart';
+import 'package:illinois/service/Analytics.dart';
+import 'package:illinois/service/Auth2.dart' as illinois;
+import 'package:illinois/ui/profile/ProfileDirectoryPage.dart';
+import 'package:illinois/ui/profile/ProfileDirectoryWidgets.dart';
+import 'package:illinois/utils/AppUtils.dart';
+import 'package:rokwire_plugin/model/auth2.directory.dart';
+import 'package:rokwire_plugin/service/auth2.dart';
+import 'package:rokwire_plugin/service/auth2.directory.dart';
+import 'package:rokwire_plugin/service/localization.dart';
+import 'package:rokwire_plugin/service/notification_service.dart';
+import 'package:rokwire_plugin/service/styles.dart';
+import 'package:rokwire_plugin/utils/utils.dart';
+
+class ProfileDirectoryAccountsList extends StatefulWidget {
+  final DirectoryAccounts contentType;
+  final DirectoryDisplayMode displayMode;
+  final ScrollController? scrollController;
+  final String? searchText;
+  final Map<String, dynamic>? filterAttributes;
+  final Set<String>? selectedAccountIds;
+  final void Function(Auth2PublicAccount, bool)? onAccountSelectionChanged;
+
+  ProfileDirectoryAccountsList(this.contentType, { super.key, this.displayMode = DirectoryDisplayMode.browse, this.scrollController,
+    this.searchText, this.filterAttributes, this.onAccountSelectionChanged, this.selectedAccountIds});
+
+  @override
+  State<StatefulWidget> createState() => ProfileDirectoryAccountsListState();
+}
+
+class ProfileDirectoryAccountsListState extends State<ProfileDirectoryAccountsList> with AutomaticKeepAliveClientMixin<ProfileDirectoryAccountsList> implements NotificationsListener  {
+
+  List<Auth2PublicAccount>? _accounts;
+  bool _loading = false;
+  bool _loadingProgress = false;
+  bool _extending = false;
+  bool _canExtend = false;
+  static const int _pageLength = 32;
+
+  String? _expandedAccountId;
+
+  String _directoryPhotoImageToken = DirectoryProfilePhotoUtils.newToken;
+  String _userPhotoImageToken = DirectoryProfilePhotoUtils.newToken;
+
+  @override
+  void initState() {
+    NotificationService().subscribe(this, [
+      illinois.Auth2.notifyProfilePictureChanged,
+      Auth2.notifyProfileChanged,
+      Auth2.notifyPrivacyChanged,
+      Auth2.notifyLoginChanged,
+    ]);
+
+    widget.scrollController?.addListener(_scrollListener);
+
+    _load();
+    super.initState();
+  }
+
+
+  @override
+  void dispose() {
+    NotificationService().unsubscribe(this);
+    widget.scrollController?.removeListener(_scrollListener);
+    super.dispose();
+  }
+
+  @override
+  void onNotification(String name, dynamic param) {
+    if (name == illinois.Auth2.notifyProfilePictureChanged) {
+      if (mounted) {
+        setState((){
+          _userPhotoImageToken = DirectoryProfilePhotoUtils.newToken;
+        });
+      }
+    }
+    else if ((name == Auth2.notifyProfileChanged) || (name == Auth2.notifyPrivacyChanged) || (name == Auth2.notifyLoginChanged)) {
+      if (mounted) {
+        setState(() {
+          _userPhotoImageToken = DirectoryProfilePhotoUtils.newToken;
+        });
+        refresh();
+      }
+    }
+  }
+
+  @override
+  bool get wantKeepAlive => true;
+
+  @override
+  Widget build(BuildContext context) {
+    super.build(context);
+    if (_loadingProgress) {
+      return _loadingContent;
+    }
+    else if (_accounts == null) {
+      return _messageContent(_failedText);
+    }
+    else if (_accounts?.isEmpty == true) {
+      return _messageContent(_emptyText);
+    }
+    else {
+      return _accountsContent;
+    }
+  }
+
+  Widget get _accountsContent {
+    List<Widget> contentList = <Widget>[];
+
+    List<Auth2PublicAccount>? accounts = _accounts;
+    if ((accounts != null) && accounts.isNotEmpty) {
+      String? directoryIndex;
+      for (Auth2PublicAccount account in accounts) {
+        String? accountDirectoryIndex = account.directoryKey;
+        if ((accountDirectoryIndex != null) && (directoryIndex != accountDirectoryIndex)) {
+          if (contentList.isNotEmpty) {
+            contentList.add(Padding(padding: EdgeInsets.only(bottom: 16), child: _sectionSplitter));
+          }
+          contentList.add(_sectionHeading(directoryIndex = accountDirectoryIndex));
+        }
+        contentList.add(_sectionSplitter);
+        contentList.add(DirectoryAccountCard(account,
+          displayMode: widget.displayMode,
+          photoImageToken: (account.id == Auth2().accountId) ? _userPhotoImageToken : _directoryPhotoImageToken,
+          expanded: (_expandedAccountId != null) && (account.id == _expandedAccountId),
+          onToggleExpanded: () => _onToggleAccountExpanded(account),
+          selected: widget.selectedAccountIds?.contains(account.id) == true,
+          onToggleSelected: (value) => _onToggleAccountSelected(account, value),
+        ));
+      }
+      if (contentList.isNotEmpty) {
+        contentList.add(Padding(padding: EdgeInsets.only(bottom: 16), child: _sectionSplitter));
+      }
+    }
+
+    if (_extending) {
+      contentList.add(_extendingIndicator);
+    }
+
+    return Column(crossAxisAlignment: CrossAxisAlignment.start, children: contentList);
+  }
+
+  void _onToggleAccountExpanded(Auth2PublicAccount account) {
+    Analytics().logSelect(target: 'Expand', source: account.id);
+    setState(() {
+      _expandedAccountId = (_expandedAccountId != account.id) ? account.id : null;
+    });
+  }
+
+  void _onToggleAccountSelected(Auth2PublicAccount account, bool value) {
+    Analytics().logSelect(target: value ? 'Select' : 'Unselect', source: account.id);
+    widget.onAccountSelectionChanged?.call(account, value);
+  }
+
+  Widget _sectionHeading(String dirEntry) =>
+    Padding(padding: EdgeInsets.zero, child:
+      Text(dirEntry, style: Styles().textStyles.getTextStyle('widget.title.small.semi_fat'),)
+    );
+
+  Widget get _sectionSplitter => Container(height: 1, color: Styles().colors.dividerLineAccent,);
+
+  Widget get _extendingIndicator => Container(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8), child:
+    Align(alignment: Alignment.center, child:
+      SizedBox(width: 24, height: 24, child:
+        DirectoryProgressWidget()
+      ),
+    ),
+  );
+
+  Widget get _loadingContent => Padding(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 64,), child:
+    Center(child:
+      SizedBox(width: 32, height: 32, child:
+        CircularProgressIndicator(color: Styles().colors.fillColorSecondary, strokeWidth: 3,)
+      )
+    )
+  );
+
+  Widget _messageContent(String message) => Padding(padding: EdgeInsets.symmetric(horizontal: 16, vertical: 64,), child:
+    Center(child:
+      Text(message, style: Styles().textStyles.getTextStyle("widget.message.dark.regular"), textAlign: TextAlign.center,)
+    )
+  );
+
+  String get _emptyText {
+    switch (widget.contentType) {
+      case DirectoryAccounts.myConnections: return Localization().getStringEx('panel.profile.directory.accounts.connections.empty.text', 'You do not have any Connections. Your connections will appear after you swap info with another ${AppTextUtils.universityLongNameMacro} student or employee.').replaceAll(AppTextUtils.universityLongNameMacro, AppTextUtils.universityLongName);
+      case DirectoryAccounts.userDirectory: return Localization().getStringEx('panel.profile.directory.accounts.directory.empty.text', 'The User Directory is empty.');
+    }
+  }
+
+  String get _failedText {
+    switch (widget.contentType) {
+      case DirectoryAccounts.myConnections: return Localization().getStringEx('panel.profile.directory.accounts.connections.failed.text', 'Failed to load Connections content.');
+      case DirectoryAccounts.userDirectory: return Localization().getStringEx('panel.profile.directory.accounts.directory.failed.text', 'Failed to load User Directory content.');
+    }
+  }
+
+  void _scrollListener() {
+    ScrollController? scrollController = widget.scrollController;
+    if ((scrollController != null) && (scrollController.offset >= scrollController.position.maxScrollExtent) && _canExtend && !_loading && !_extending) {
+      _extend();
+    }
+  }
+
+  Future<void> _load({ int limit = _pageLength, bool silent = false }) async {
+    if (!_loading) {
+      setStateIfMounted(() {
+        _loading = true;
+        _loadingProgress = !silent;
+        _extending = false;
+      });
+
+      List<Auth2PublicAccount>? accounts = await Auth2().loadDirectoryAccounts(
+        search: StringUtils.ensureEmpty(widget.searchText),
+        attriutes: widget.filterAttributes,
+        limit: limit,
+      );
+
+      setStateIfMounted(() {
+        _loading = false;
+        _loadingProgress = false;
+        if (accounts != null) {
+          _accounts = List.from(accounts);
+          _canExtend = (accounts.length >= limit);
+        }
+        else if (!silent) {
+          _accounts = null;
+          _canExtend = false;
+        }
+      });
+    }
+  }
+
+  Future<void> refresh() =>
+    _load(limit: max(_accountsCount, _pageLength), silent: true);
+
+  Future<void> _extend() async {
+    if (!_loading && !_extending) {
+      setStateIfMounted(() {
+        _extending = true;
+      });
+
+      List<Auth2PublicAccount>? accounts = await Auth2().loadDirectoryAccounts(
+        search: StringUtils.ensureEmpty(widget.searchText),
+        attriutes: widget.filterAttributes,
+        offset: _accountsCount,
+        limit: _pageLength
+      );
+
+      if (mounted && _extending && !_loading) {
+        setState(() {
+          if (accounts != null) {
+            if (_accounts != null) {
+              _accounts?.addAll(accounts);
+            }
+            else {
+              _accounts = List.from(accounts);
+            }
+
+            _canExtend = (accounts.length >= _pageLength);
+          }
+          _extending = false;
+        });
+      }
+    }
+  }
+
+  int get _accountsCount => _accounts?.length ?? 0;
+}
+
+extension _Auth2PublicAccountUtils on Auth2PublicAccount {
+  String? get directoryKey => (profile?.lastName?.isNotEmpty == true) ?
+    profile?.lastName?.substring(0, 1).toUpperCase() : null;
+}
